Create A Password Validiator
This is a password validiator i made followoing this steps.
I made the string a list,  after importing the module, so i could iltrate through each of them.  
Secondly, i created each function to validiate each properties of a strong Password( must be more than 8 in charater,
must contain small and capital letter including punctuations.)
i created a function to check if all the properties of each functions are met and gives report on which parameter are not met.
The main function checks if all requirments are met and gives a postive result, if the password isn't strong it asks the user to input a better password.